pyproject.toml
[project]
name = "gitlog-summary"
version = "0.1.0"
description = "Aggregate all git commits across multiple repositories for a given day and generate a summary."
authors = [
    {name = "mcgillij",email = "mcgillivray.jason@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "pygithub (>=2.6.1,<3.0.0)",
    "click (>=8.2.1,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
gitlog-summary = "gitlog_summary.cli:main"
